+++ ml-agents/mlagents/trainers/tests/check_env_trains.py
@@ -45,7 +45,7 @@ def default_reward_processor(rewards, last_n_rewards=5):
     rewards_to_use = rewards[-last_n_rewards:]
     # For debugging tests
     print(f"Last {last_n_rewards} rewards:", rewards_to_use)
-    return np.array(rewards[-last_n_rewards:], dtype=np.float32).mean()
+    return np.array(rewards[-last_n_rewards:], dtype=float).mean()
 
 
 def check_environment_trains(
